blog and core values sections
