+++ src/components/TestimonyTile.js
@@ -45,8 +45,6 @@ const TestimonyTile = ({
                         {date}
                     </Text>
                 </Col>
-            </Row>
-            <Row>
                 <Col md={12}>
                     <Title blue>
                         {headline}
